First package factored out of admin. Not exactly sure where it should go, so placement needs to be revisited later
