+++ router/admin/anycred/creds.go
@@ -1,4 +1,4 @@
-package admin
+package anycred
 
 import (
 	"context"
@@ -6,21 +6,34 @@ import (
 
 	"github.com/golang/protobuf/ptypes/empty"
 	"github.com/level11consulting/ocelot/models/pb"
+	"github.com/level11consulting/ocelot/storage"
+	"github.com/level11consulting/ocelot/server/config"
 
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
 
-func (g *OcelotServerAPI) CheckAnyCredExists(ctx context.Context, creds pb.OcyCredder) (*pb.Exists, error) {
-	exists, err := g.DeprecatedHandler.Storage.CredExists(creds)
+type AnyCred interface {
+	CheckAnyCredExists(ctx context.Context, creds pb.OcyCredder) (*pb.Exists, error)
+	UpdateAnyCred(ctx context.Context, creds pb.OcyCredder) (*empty.Empty, error)
+}
+
+type AnyCredAPI struct {
+	AnyCred
+	RemoteConfig   config.CVRemoteConfig
+	Storage        storage.OcelotStorage
+}
+
+func (g *AnyCredAPI) CheckAnyCredExists(ctx context.Context, creds pb.OcyCredder) (*pb.Exists, error) {
+	exists, err := g.Storage.CredExists(creds)
 	if err != nil {
 		return nil, status.Errorf(codes.Unavailable, "Unable to reach cred table to check if cred %s/%s/%s exists. Error: %s", creds.GetAcctName(), creds.GetSubType().String(), creds.GetIdentifier(), err.Error())
 	}
 	return &pb.Exists{Exists: exists}, nil
 }
 
-func (g *OcelotServerAPI) UpdateAnyCred(ctx context.Context, creds pb.OcyCredder) (*empty.Empty, error) {
-	if err := g.DeprecatedHandler.RemoteConfig.UpdateCreds(g.DeprecatedHandler.Storage, creds); err != nil {
+func (g *AnyCredAPI) UpdateAnyCred(ctx context.Context, creds pb.OcyCredder) (*empty.Empty, error) {
+	if err := g.RemoteConfig.UpdateCreds(g.Storage, creds); err != nil {
 		if _, ok := err.(*pb.ValidationErr); ok {
 			return &empty.Empty{}, status.Errorf(codes.InvalidArgument, "%s cred failed validation. Errors are: %s", creds.GetSubType().Parent(), err.Error())
 		}
@@ -29,7 +42,7 @@ func (g *OcelotServerAPI) UpdateAnyCred(ctx context.Context, creds pb.OcyCredder
 	return &empty.Empty{}, nil
 }
 
-func (g *OcelotServerAPI) DeleteAnyCred(ctx context.Context, creds pb.OcyCredder, parentType pb.CredType) (*empty.Empty, error) {
+func (g *AnyCredAPI) DeleteAnyCred(ctx context.Context, creds pb.OcyCredder, parentType pb.CredType) (*empty.Empty, error) {
 	// make sure we have all the fields we need to be able to accurately delete the credential.
 	// try to intelligently deduce what subType teh cred is, but error out if that isn't possible
 	empti := &empty.Empty{}
@@ -48,11 +61,11 @@ func (g *OcelotServerAPI) DeleteAnyCred(ctx context.Context, creds pb.OcyCredder
 		return empti, status.Error(codes.InvalidArgument, errmsg)
 	}
 
-	if exists, _ := g.DeprecatedHandler.Storage.CredExists(creds); !exists {
+	if exists, _ := g.Storage.CredExists(creds); !exists {
 		return empti, status.Error(codes.NotFound, "not found")
 	}
 
-	if err := g.DeprecatedHandler.RemoteConfig.DeleteCred(g.DeprecatedHandler.Storage, creds); err != nil {
+	if err := g.RemoteConfig.DeleteCred(g.Storage, creds); err != nil {
 		return empti, status.Error(codes.Internal, err.Error())
 	}
 	return empti, nil
