+++ router/admin/anycred/creds_test.go
@@ -1,4 +1,4 @@
-package admin
+package anycred
 
 import (
 	"context"
